+++ app/Exceptions/Passkey/PasskeyException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Exceptions\Passkey;
+
+use RuntimeException;
+
+class PasskeyException extends RuntimeException
+{
+}